+++ src/ConvertToSSM/Util.ts
@@ -70,6 +70,28 @@ export function validate(languageId: string): boolean {
   }
 }
 
+export function validateScriptType(languageId: string): boolean {
+  if (!(['powershell', 'shellscript'].includes(languageId))) {
+    vscode.window.showErrorMessage('Only PowerShell and Shell scripts supported :(');
+    return false;
+  }
+  else {
+    return true;
+  }
+}
+
+
+export function validateDocumentType(languageId: string): boolean {
+  if (!(['json', 'yaml'].includes(languageId))) {
+    vscode.window.showErrorMessage('Only JSON and YAML SSM documents supported :(');
+    return false;
+  }
+  else {
+    return true;
+  }
+}
+
+
 export interface DocumentStructure {
   schemaVersion: string;
   description: string;